# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Donation:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      receiver_id:
        type: integer
      sender_id:
        type: integer
      video_id:
        type: integer
    type: object
...
  dto.CreateDonationRequest:
    properties:
      amount:
        type: number
      message:
        type: string
      receiver_id:
        type: integer
      recipient_id:
        type: integer
      video_id:
        type: integer
    type: object
...
  httpdelivery.CommonError:
    properties:
      error:
        type: string
    type: object
  httpdelivery.CommonResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: analogyoutube@example.com
    name: AnalogYouTube Backend
  description: |-
    API для AnalogYouTube: пользователи, видео, чаты, донаты и остальное.
    Для закрытых запросов нажмите Authorize и вставьте: Bearer <token>.
  title: AnalogYouTube API
  version: "1.0"
paths:
...
  /api/donations:
    post:
      consumes:
      - application/json
      description: Сохраняет донат без настоящей оплаты.
      parameters:
      - description: Данные доната
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDonationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Donation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Создать донат
      tags:
      - Donations
  /api/donations/received:
    get:
      description: Донаты, которые получили вы.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Donation'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Мои полученные донаты
      tags:
      - Donations
  /api/donations/sent:
    get:
      description: Донаты, которые отправили вы.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Donation'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Мои отправленные донаты
      tags:
      - Donations
...
  /api/users/{id}/donations:
    get:
      description: Публичный список полученных донатов.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Donation'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Донаты пользователя
      tags:
      - Donations
...
  /api/videos/{id}/like:
    delete:
      description: Удаляет мой лайк с видео.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Убрать лайк
      tags:
      - Likes
    post:
      description: Лайк от текущего пользователя.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Поставить лайк
      tags:
      - Likes
  /api/videos/{id}/liked:
    get:
      description: Показывает, лайкнул ли я видео.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Мой лайк
      tags:
      - Likes
  /api/videos/{id}/likes/count:
    get:
      description: Сколько лайков у видео.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Количество лайков
      tags:
      - Likes
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"